folder structure and creating the database
