+++ src/lib.rs
@@ -1,25 +1,3 @@
-use actix_web::{ 
-    web,
-    App, 
-    HttpResponse,
-    HttpServer, 
-    dev::Server
-};
-use std::net::TcpListener;
-
-pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(|| {
-        App::new()
-            .route("/health_check", web::get().to(health_check))
-    })
-    .listen(listener)?
-    .run();
-
-    Ok(server)
-}
-
-async fn health_check() -> HttpResponse {
-    HttpResponse::Ok().finish()
-}
-
-
+pub mod configuration;
+pub mod routes;
+pub mod startup;
