+++ src/routes/health_check.rs
@@ -1,0 +1,6 @@
+use actix_web::HttpResponse;
+
+// health check
+pub async fn health_check() -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
