+++ src/routes/subscriptions.rs
@@ -1,0 +1,15 @@
+use actix_web::{ 
+    web,
+    HttpResponse
+};
+
+// suscribe
+#[derive(serde::Deserialize)]
+pub struct FormData {
+    email: String,
+    name: String
+}
+
+pub async fn subscribe(_form: web::Form<FormData>) -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
