+++ tests/health_check.rs
@@ -69,7 +69,7 @@ async fn subscribe_returns_400() {
 fn spawn_app() -> String {
     let listener = TcpListener::bind("127.0.0.1:0").expect("Failed to bind random port");
     let port = listener.local_addr().unwrap().port();
-    let server = email_newsletter::run(listener).expect("Failed to bind address");
+    let server = email_newsletter::startup::run(listener).expect("Failed to bind address");
     let _ = tokio::spawn(server);
     //We return the application address to the caller
     format!("http://127.0.0.1:{}", port)
